# docs/swagger.yaml
basePath: /api/v1/
definitions:
...
  domain.City:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  domain.GetAllNotificationsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.Notification'
        type: array
      page-info:
        $ref: '#/definitions/domain.PaginationPage'
    type: object
...
  domain.GetterList:
    properties:
      id:
        type: integer
      notification_id:
        type: integer
      user_id:
        type: integer
    type: object
  domain.Notification:
    properties:
      date:
        type: number
      getters:
        type: string
      id:
        type: integer
      ids:
        items:
          type: integer
        type: array
      link:
        type: string
      partner_id:
        type: integer
      reference:
        type: string
      status:
        type: string
      text:
        type: string
      time:
        type: number
      title:
        type: string
      users:
        $ref: '#/definitions/domain.GetterList'
    required:
    - text
    - title
    type: object
  domain.PaginationPage:
    properties:
      count:
        type: integer
      page:
        type: integer
      pages:
        type: integer
    type: object
...
  v1.City:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  v1.Notification:
    properties:
      date:
        type: number
      link:
        type: string
      partner_id:
        type: integer
      reference:
        type: string
      text:
        type: string
      time:
        type: number
      title:
        type: string
    required:
    - text
    - title
    type: object
  v1.NotificationForUser:
    properties:
      link:
        type: string
      text:
        type: string
      title:
        type: string
      user_ids:
        items:
          type: integer
        type: array
    required:
    - text
    - title
    - user_ids
    type: object
...
  v1.RegistrationTokenInput:
    properties:
      registration_token:
        type: string
    required:
    - registration_token
    type: object
...
  v1.idResponse:
    properties:
      id: {}
    type: object
  v1.okResponse:
    properties:
      message:
        type: string
    type: object
...
  v1.response:
    properties:
      detail:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: API Server for  A Hundred to Five
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: A Hundred to Five
  version: "2.0"
paths:
...
  /city/{id}:
    delete:
      consumes:
      - application/json
      description: delete city
      parameters:
      - description: city id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.okResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - User_Auth: []
      tags:
      - city
    get:
      consumes:
      - application/json
      description: get city by id
      operationId: get-city-by-id
      parameters:
      - description: city id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.City'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      tags:
      - city
    put:
      consumes:
      - application/json
      description: update  city
      parameters:
      - description: city id
        in: path
        name: id
        required: true
        type: string
      - description: city
        in: body
        name: input
        schema:
          $ref: '#/definitions/v1.City'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.okResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - User_Auth: []
      tags:
      - city
...
  /notification:
    get:
      consumes:
      - application/json
      description: get all notifications
      operationId: get-all-notifications
      parameters:
      - in: query
        minimum: 0
        name: limit
        type: integer
      - in: query
        minimum: 0
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetAllNotificationsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      tags:
      - notification
  /notification/{id}:
    delete:
      consumes:
      - application/json
      description: delete notification
      parameters:
      - description: notification id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.okResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - User_Auth: []
      tags:
      - notification
    get:
      consumes:
      - application/json
      description: get notification by id
      operationId: get-notification-by-id
      parameters:
      - description: notification id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Notification'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      tags:
      - notification
  /notification/all:
    post:
      consumes:
      - application/json
      description: create notification for all users
      operationId: createNotyForAllUsers
      parameters:
      - description: notification
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/v1.Notification'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.idResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - User_Auth: []
      tags:
      - notification
  /notification/response:
    post:
      consumes:
      - application/json
      description: accept registration token for push notification
      operationId: notification-responsible
      parameters:
      - description: registration token
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.RegistrationTokenInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.okResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      tags:
      - notification
  /notification/user:
    post:
      consumes:
      - application/json
      description: create notification for specific user
      operationId: createNotyForSpecificUser
      parameters:
      - description: notification
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/v1.NotificationForUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.idResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - User_Auth: []
      tags:
      - notification
...
securityDefinitions:
  User_Auth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"